Http server vendoring.
This commit adds the `shs` library as a vendored dependency of lacord,
this makes a lot of the boilerplate in `lua/lacord/outgoing-webhook-server-1.lua'
redundant. Currently the two versions of that module are gated by the
unstable flag, however the next release will bring these changes forward.

New module `lua/lacord/outgoing-webhook-server-2.lua':
  - Rewritten outgoing webhook.