Deprecate api:batch_edit_application_command_permissions.
Discord's new permission system has gone live and thus the aforementioned
api method is now disabled. In the absence of --unstable and --deprecated
the method will emit a warning and return values which signal an error.
When --unstable is set this method will no longer be exported. When
--deprecated is set without --unstable the method will remain unchanged.

Changes I made to `lua/lacord/api.lua':
  - See summary above.